pyproject.toml
[project]
name = "databricks-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "databricks-sdk>=0.73.0",
    "mcp[cli]>=1.21.1",
    "requests>=2.32.5",
]
